RegionMatchesTest
Two string literal inputs are compared using the String class regionMatches method. The program also takes an integer starting index input and an integer number of characters to compared input, which are placed in arguments in the regionMatches method. The case of the String literals is ignored and the comparison is done.
JOptionPane static methods showInputDialog and showMessageDialog are used to provide the UI for this program.