Airline Reservation System
The Airline Reservation System uses 10 element, one dimensional array to store Boolean values used to indicate whether the seat is occupied or is free.
First class seats are from array index 0 - 4, while economy seats are from array index 5 - 9. Both First-class and economy have five seats.
When a seat in any section is occupied, the section chosen and the seat occupied is displayed for the user to see. A seat that is already occupied cannot be occupied again, by another customer. When the First-class seats are occupied, the user is given an option as to whether they would like to occupy economy seats.
The same is done for economy seats as well. When economy seats are occupied, the user is asked whether they would like to occupy First-class seats.
The array is displayed after each questioning to show the seats occupied and unoccupied in both sections.
JOptionPane's static methods showInputDialog() and showMessageDialog() is used to provide the UI for this code.