Pig Latin
This code encodes English-language phrases into pig Latin.  To form a pig Latin phrase from an English-language phrase, tokenize the phrase into words with String method split. To translate each English word into a pig Latin word, place the first letter of the English word at the end of the word and add the letters “ay.” Thus, the word “jump” becomes “umpjay”, the word “the” becomes “hetay,” and the word “computer” becomes “omputercay".
JOptionPane's static method showInputDialog is used to provide the UI for user input and the output is displayed on the command line.