Tokenizing Mobile Numbers
This application takes a telephone number input as a String in the form (555) 555-5555. The application uses String method spilt to extract the area code and the first three digits of the phone number as a token and the last four digits of the phone number as a token. The area code is printed out separately from the phone number. The UI is achieved using JOptionPane's static methods showMessageDialog and showInputDialog.