Search String 1
This application takes a String sentence input and a char value input. The char value input is searched for in the String sentence and the number of occurrences of the character in the String sentence is noted and stored in a counter variable. The number of occurrences is displayed at the end of the application.
UI is provided by JOptionPane's static methods, showInputDialog and showMessageDialog.