Search String 2
This application takes a line of text input and uses String method indexOf to determine the total number of occurrences of each letter of the alphabet in the text. Uppercase and lowercase letters are counted together. The total of each letter is stored in an array and is printed out after the totals have been determined.